+++ parser_library/include/debugger.h
@@ -49,13 +49,32 @@ class breakpoints_t
 
 public:
     breakpoints_t();
+    breakpoints_t(const breakpoints_t&) = delete;
     breakpoints_t(breakpoints_t&&) noexcept;
+    breakpoints_t& operator=(const breakpoints_t&) = delete;
     breakpoints_t& operator=(breakpoints_t&&) & noexcept;
     ~breakpoints_t();
 
     [[nodiscard]] const breakpoint* begin() const;
     [[nodiscard]] const breakpoint* end() const;
     [[nodiscard]] std::size_t size() const;
+};
+
+class evaluated_expression_t
+{
+    class impl;
+    impl* pimpl;
+    friend class debugger;
+
+public:
+    evaluated_expression_t();
+    evaluated_expression_t(evaluated_expression_t&&) noexcept;
+    evaluated_expression_t& operator=(evaluated_expression_t&&) & noexcept;
+    ~evaluated_expression_t();
+
+    [[nodiscard]] sequence<char> result() const noexcept;
+    [[nodiscard]] bool is_error() const noexcept;
+    [[nodiscard]] std::size_t var_ref() const noexcept;
 };
 
 // Implements DAP for macro tracing. Starts analyzer in a separate thread
@@ -103,6 +122,7 @@ public:
     stack_frames_t stack_frames() const;
     scopes_t scopes(frame_id_t frame_id) const;
     variables_t variables(var_reference_t var_ref) const;
+    evaluated_expression_t evaluate(sequence<char> expr, frame_id_t id = -1);
 
     void analysis_step(const std::atomic<unsigned char>* yield_indicator);
 };