+++ parser_library/src/debugging/debugger.cpp
@@ -27,16 +27,29 @@
 
 #include "analyzer.h"
 #include "context/hlasm_context.h"
+#include "context/ordinary_assembly/ordinary_assembly_dependency_solver.h"
 #include "context/variables/system_variable.h"
 #include "debug_lib_provider.h"
 #include "debug_types.h"
 #include "debugger_configuration.h"
+#include "diagnostic.h"
+#include "diagnostic_consumer.h"
+#include "expressions/evaluation_context.h"
+#include "lexing/input_source.h"
+#include "lexing/tools.h"
+#include "library_info_transitional.h"
 #include "macro_param_variable.h"
 #include "ordinary_symbol_variable.h"
+#include "parsing/parser_impl.h"
+#include "processing/op_code.h"
 #include "processing/statement.h"
 #include "processing/statement_analyzers/statement_analyzer.h"
+#include "processing/statement_fields_parser.h"
+#include "semantics/operand_impls.h"
 #include "set_symbol_variable.h"
 #include "utils/async_busy_wait.h"
+#include "utils/factory.h"
+#include "utils/string_operations.h"
 #include "utils/task.h"
 #include "variable.h"
 #include "workspace_manager.h"
@@ -78,6 +91,53 @@ const breakpoint* breakpoints_t::begin() const { return pimpl->m_breakpoints.dat
 const breakpoint* breakpoints_t::end() const { return pimpl->m_breakpoints.data() + pimpl->m_breakpoints.size(); }
 std::size_t breakpoints_t::size() const { return pimpl->m_breakpoints.size(); }
 
+class evaluated_expression_t::impl
+{
+    friend class debugger;
+    friend class evaluated_expression_t;
+
+    std::string result;
+    bool error = false;
+    std::size_t var_ref = 0;
+
+    void set_value(std::string s) noexcept
+    {
+        result = std::move(s);
+        error = false;
+    }
+
+    void set_error(std::string s) noexcept
+    {
+        result = std::move(s);
+        error = true;
+    }
+};
+
+evaluated_expression_t::evaluated_expression_t()
+    : pimpl(new impl())
+{}
+
+evaluated_expression_t::evaluated_expression_t(evaluated_expression_t&& o) noexcept
+    : pimpl(std::exchange(o.pimpl, nullptr))
+{}
+
+evaluated_expression_t& evaluated_expression_t::operator=(evaluated_expression_t&& o) & noexcept
+{
+    auto tmp(std::move(o));
+    std::swap(pimpl, tmp.pimpl);
+    return *this;
+}
+
+evaluated_expression_t::~evaluated_expression_t()
+{
+    if (pimpl)
+        delete pimpl;
+}
+
+[[nodiscard]] sequence<char> evaluated_expression_t::result() const noexcept { return sequence<char>(pimpl->result); }
+[[nodiscard]] bool evaluated_expression_t::is_error() const noexcept { return pimpl->error; }
+[[nodiscard]] std::size_t evaluated_expression_t::var_ref() const noexcept { return pimpl->var_ref; }
+
 // Implements DAP for macro tracing. Starts analyzer in a separate thread
 // then controls the flow of analyzer by implementing processing_tracer
 // interface.
@@ -99,10 +159,11 @@ class debugger::impl final : public processing::statement_analyzer
 
     // Debugging information retrieval
     context::hlasm_context* ctx_ = nullptr;
+    parse_lib_provider* lib_provider_ = nullptr;
     std::string opencode_source_uri_;
     std::vector<stack_frame> stack_frames_;
     std::vector<scope> scopes_;
-    std::unordered_map<frame_id_t, context::hlasm_context::sysvar_map> last_system_variables_;
+    std::unordered_map<frame_id_t, context::system_variable_map> last_system_variables_;
 
     std::unordered_map<size_t, variable_store> variables_;
     size_t next_var_ref_ = 1;
@@ -155,6 +216,7 @@ class debugger::impl final : public processing::statement_analyzer
         a.register_stmt_analyzer(this);
 
         ctx_ = a.context().hlasm_ctx.get();
+        lib_provider_ = &debug_provider;
 
         co_await a.co_analyze();
     }
@@ -443,6 +505,287 @@ public:
         return it->second;
     }
 
+    struct error_collector final : diagnostic_op_consumer
+    {
+        std::string& msg;
+        explicit error_collector(std::string& msg)
+            : msg(msg)
+        {}
+
+        void add_diagnostic(diagnostic_op d) const override
+        {
+            if (d.severity != diagnostic_severity::error)
+                return;
+            if (!msg.empty())
+                msg.push_back('\n');
+            msg.append(d.code).append(": ").append(d.message);
+        }
+    };
+
+    void evaluate_expression(evaluated_expression_t::impl& result,
+        const expressions::ca_expression& expr,
+        const context::code_scope& scope,
+        const context::system_variable_map& sysvars) const
+    {
+        std::string error_msg;
+        error_collector diags(error_msg);
+        library_info_transitional lib_info(*lib_provider_);
+
+        auto eval = expr.evaluate({ *ctx_, lib_info, diags, scope, sysvars });
+
+        if (!error_msg.empty())
+            return result.set_error(std::move(error_msg));
+
+        using enum context::SET_t_enum;
+        switch (eval.type())
+        {
+            case UNDEF_TYPE:
+                result.set_value("<UNDEFINED>");
+                break;
+
+            case A_TYPE:
+                result.set_value(std::to_string(eval.access_a()));
+                break;
+
+            case B_TYPE:
+                result.set_value(eval.access_b() ? "TRUE" : "FALSE");
+                break;
+
+            case C_TYPE:
+                result.set_value(std::move(eval.access_c()));
+                break;
+        }
+    }
+
+    static std::string to_string(const context::address& reloc)
+    {
+        std::string text;
+
+        auto bases = std::vector<context::address::base_entry>(reloc.bases().begin(), reloc.bases().end());
+        std::sort(bases.begin(), bases.end(), [](const auto& l, const auto& r) {
+            return l.first.owner->name < r.first.owner->name;
+        });
+        bool first = true;
+        for (const auto& [base, d] : bases)
+        {
+            if (base.owner->name.empty() || d == 0)
+                continue;
+
+            bool was_first = std::exchange(first, false);
+            if (d < 0)
+                text.append(was_first ? "-" : " - ");
+            else if (!was_first)
+                text.append(" + ");
+
+            if (d != 1 && d != -1)
+                text.append(std::to_string(d < 0 ? -(unsigned)d : (unsigned)d)).append("*");
+
+            if (!base.qualifier.empty())
+                text.append(base.qualifier.to_string_view()).append(".");
+            text.append(base.owner->name.to_string_view());
+        }
+        if (!first)
+            text.append(" + ");
+
+        text.append("X'");
+        uint32_t offset = reloc.offset();
+        size_t len = text.size();
+        do
+        {
+            text.push_back("0123456789ABCDEF"[offset & 0xf]);
+            offset >>= 4;
+
+        } while (offset);
+        std::reverse(text.begin() + len, text.end());
+        text.push_back('\'');
+
+        return text;
+    }
+
+    void evaluate_expression(evaluated_expression_t::impl& result, const expressions::mach_expression& expr) const
+    {
+        std::string error_msg;
+        error_collector diags(error_msg);
+        library_info_transitional lib_info(*lib_provider_);
+        context::ordinary_assembly_dependency_solver dep_solver(ctx_->ord_ctx,
+            // do not introduce a private section by accident
+            // TODO: the alignment will be wrong
+            //       - it is instruction and argument dependent which is not available yet
+            ctx_->ord_ctx.current_section() ? ctx_->ord_ctx.align(context::no_align, lib_info) : context::address(),
+            lib_info);
+
+        if (auto d = expr.get_dependencies(dep_solver); d.has_error)
+            return result.set_error("Expression cannot be evaluated");
+        else if (d.contains_dependencies())
+            return result.set_error("Expression has unresolved dependencies");
+
+        auto eval = expr.evaluate(dep_solver, diags).ignore_qualification();
+
+        if (!error_msg.empty())
+            return result.set_error(std::move(error_msg));
+
+        using enum context::symbol_value_kind;
+        switch (eval.value_kind())
+        {
+            case UNDEF:
+                result.set_value("<UNDEFINED>");
+                break;
+
+            case ABS:
+                result.set_value(std::to_string(eval.get_abs()));
+                break;
+
+            case RELOC:
+                result.set_value(to_string(eval.get_reloc()));
+                break;
+        }
+    }
+
+    static constexpr const processing::processing_status mach_status = {
+        processing::processing_format(processing::processing_kind::ORDINARY, processing::processing_form::MACH),
+        processing::op_code()
+    };
+    static constexpr const processing::processing_status seta_status = {
+        processing::processing_format(processing::processing_kind::ORDINARY, processing::processing_form::CA),
+        processing::op_code(context::id_storage::well_known::SETA, context::instruction_type::CA, nullptr)
+    };
+    static constexpr const processing::processing_status setb_status = {
+        processing::processing_format(processing::processing_kind::ORDINARY, processing::processing_form::CA),
+        processing::op_code(context::id_storage::well_known::SETB, context::instruction_type::CA, nullptr)
+    };
+    static constexpr const processing::processing_status setc_status = {
+        processing::processing_format(processing::processing_kind::ORDINARY, processing::processing_form::CA),
+        processing::op_code(context::id_storage::well_known::SETC, context::instruction_type::CA, nullptr)
+    };
+
+    void evaluate_exact_match(evaluated_expression_t::impl& result,
+        std::string_view var_name,
+        const context::code_scope& scope,
+        const context::system_variable_map& sysvars)
+    {
+        debugging::variable_ptr var;
+
+        if (scope.is_in_macro())
+        {
+            for (const auto& [name, value] : scope.this_macro->named_params)
+            {
+                if (name.to_string_view() != var_name)
+                    continue;
+                var = std::make_unique<macro_param_variable>(*value, std::vector<context::A_t> {});
+                break;
+            }
+        }
+
+        if (!var)
+        {
+            for (const auto& [name, value] : scope.variables)
+            {
+                if (name.to_string_view() != var_name)
+                    continue;
+                var = std::make_unique<set_symbol_variable>(*value.ref);
+                break;
+            }
+        }
+
+        if (!var)
+        {
+            for (const auto& [name, value] : sysvars)
+            {
+                if (name.to_string_view() != var_name)
+                    continue;
+                var = std::make_unique<macro_param_variable>(*value.first, std::vector<context::A_t> {});
+                break;
+            }
+        }
+
+        if (!var)
+            return result.set_error("Variable not found");
+
+        result.set_value(var->get_value());
+        if (!var->is_scalar())
+        {
+            result.var_ref = add_variable(var->values());
+        }
+    }
+
+    evaluated_expression_t evaluate(std::string_view expr, frame_id_t frame_id)
+    {
+        evaluated_expression_t result;
+
+        if (debug_ended_ || expr.empty())
+            return result;
+
+        if (frame_id == (size_t)-1)
+            frame_id = proc_stack_.size() - 1;
+
+        if (frame_id >= proc_stack_.size())
+        {
+            result.pimpl->set_error("Invalid frame id");
+            return result;
+        }
+        const auto& scope = proc_stack_[frame_id].scope;
+        auto [sysvars, _] = last_system_variables_.try_emplace(
+            frame_id, utils::factory([this, &scope]() { return ctx_->get_system_variables(scope); }));
+
+        if (expr.starts_with("&") && lexing::is_valid_symbol_name(expr.substr(1)))
+        {
+            evaluate_exact_match(*result.pimpl, utils::to_upper_copy(expr.substr(1)), scope, sysvars->second);
+            return result;
+        }
+
+        static constexpr const auto stringy_attribute = [](std::string_view s) {
+            return s.starts_with("T'") || s.starts_with("O'") || s.starts_with("t'") || s.starts_with("o'");
+        };
+
+        const auto& status = [&expr]() -> const auto&
+        {
+            if (const auto pos = expr.find("&"); pos == (size_t)-1)
+                return mach_status;
+            else if (expr.front() == '(' && expr.back() == ')')
+                return setb_status;
+            else if (expr.front() == '\'' || stringy_attribute(expr))
+                return setc_status;
+            else
+                return seta_status;
+        }
+        ();
+
+        std::string error_msg;
+        error_collector diags(error_msg);
+
+        auto p = parsing::parser_holder::create(nullptr, ctx_, nullptr, false);
+        p->prepare_parser(expr, ctx_, &diags, semantics::range_provider(), range(), 1, status, true);
+
+        semantics::operand_ptr op =
+            status.first.form == processing::processing_form::CA ? p->ca_op_expr() : p->operand_mach();
+
+        static constexpr auto ca_expr = [](const semantics::operand_ptr& o) -> const expressions::ca_expression* {
+            if (auto* ca_op = o->access_ca())
+                return ca_op->access_expr()->expression.get();
+            return nullptr;
+        };
+        static constexpr auto mach_expr = [](const semantics::operand_ptr& o) -> const expressions::mach_expression* {
+            if (auto* mach_op = o->access_mach())
+                return mach_op->access_expr()->expression.get();
+            return nullptr;
+        };
+
+        if (p->input->LA(1) != (size_t)-1)
+            result.pimpl->set_error("Invalid expression");
+        else if (!error_msg.empty())
+            result.pimpl->set_error(std::move(error_msg));
+        else if (!op)
+            result.pimpl->set_error("Single expression expected");
+        else if (const auto* caop = ca_expr(op))
+            evaluate_expression(*result.pimpl, *caop, scope, sysvars->second);
+        else if (const auto* mop = mach_expr(op))
+            evaluate_expression(*result.pimpl, *mop);
+        else
+            result.pimpl->set_error("Unexpected operand format");
+
+        return result;
+    }
+
     void breakpoints(const utils::resource::resource_location& source, std::vector<breakpoint> bps)
     {
         breakpoints_[source] = std::move(bps);
@@ -523,6 +866,11 @@ variables_t debugger::variables(var_reference_t var_ref) const
 {
     const auto& v = pimpl->variables(var_ref);
     return variables_t(&v, v.variables.size());
+}
+
+evaluated_expression_t debugger::evaluate(sequence<char> expr, frame_id_t id)
+{
+    return pimpl->evaluate(std::string_view(expr), id);
 }
 
 } // namespace hlasm_plugin::parser_library::debugging