+++ parser_library/src/processing/op_code.h
@@ -30,10 +30,10 @@ namespace hlasm_plugin::parser_library::processing {
 // structure holding resolved operation code of the instruction (solving OPSYNs and so on)
 struct op_code
 {
-    op_code()
+    constexpr op_code()
         : type(context::instruction_type::UNDEF)
     {}
-    op_code(context::id_index value, context::instruction_type type, const context::macro_definition* mac_def)
+    constexpr op_code(context::id_index value, context::instruction_type type, const context::macro_definition* mac_def)
         : value(value)
         , mac_def(mac_def)
         , type(type)