+++ parser_library/src/processing/processing_format.h
@@ -48,7 +48,7 @@ enum class operand_occurrence : uint8_t
 // structure respresenting in which fashion should be statement processed
 struct processing_format
 {
-    processing_format(
+    constexpr processing_format(
         processing_kind kind, processing_form form, operand_occurrence occurrence = operand_occurrence::PRESENT)
         : kind(kind)
         , form(form)